Move code from pattern matcher
